fix pdf generation tool after review
