+++ ada_backend/services/agent_runner_service.py
@@ -20,6 +20,7 @@ from ada_backend.repositories.organization_repository import get_organization_se
 from engine.graph_runner.runnable import Runnable
 from engine.trace.trace_context import get_trace_manager
 from engine.trace.span_context import set_tracing_span
+from engine.agent.pdf_generation_tool import delete_pdf_file_if_exists
 
 
 def get_organization_llm_providers(session: Session, organization_id: UUID) -> list[str]:
@@ -133,6 +134,7 @@ async def run_agent(
             input_data,
             is_root_execution=True,
         )
+        delete_pdf_file_if_exists(agent_output)
     except Exception as e:
         raise ValueError(f"Error running agent: {str(e)}") from e
     return ChatResponse(
